+++ src/dagagent/planner/cookbooks/__init__.py
@@ -1,0 +1,1 @@
+"""Bundled planner cookbook recipes."""